requirements.txt
asyncio-nats-client==0.11.4
pkg-resources==0.0.0
